requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
tomli>=2.0.1; python_version < "3.11"
requests~=2.31.0
azure-cognitiveservices-speech~=1.34.1
paho-mqtt~=1.6.1
pycsspeechtts~=1.0.8
apscheduler>=3.10
SQLAlchemy>=2.0
PyJWT>=2.8.0
bcrypt>=4.0.0